GDv2/tests: Restrict char set of profiling attributes (fixes #90)
For GeekoDoc version 1 and 2:

Profiling attributes (like os, arch, ...) only make
sense if they are separated by semicolon. Spaces or
commas are not allowed and make the profiling algorithm
fail. Additionally, only lowercase letters,
underscores, digits, and periods are allowed.

This commit introduces a regex to cover this use case.
NB: Relax NG regexes are implicitly at the head and tail
of attribute values, apparently, `^` and `$` are interpreted
literally, making the regex fail if these characters do
not appear literally in the value.

Co-authored-by: Stefan Knorr <sknorr@suse.de>