GeekoDoc v2: remove `@format` from `mediaobject`
On the most recent Tools Impl group call, I asked whether it would be
useful to add new image formats like DITAA to the @format attribute.
* Frank responded that this attribute is not used by DAPS's image
  conversion magic at all. That functionality only looks at the file
  name.
* Toms mentioned that this attribute is essentially only useful when the
  file extension does not match the file type. (Which is a use case
  explicitly disallowed in DAPS.)

Both were a bit hesitant to just remove the attribute because getting
rid of it within existing documents is work (there are 600-odd
occurrences of it in doc-sle's main). However, I still think it's a
good idea to get rid of a useless and possibly confusing attribute
eventually, even if that causes some short-term grief.
We also have to option to roll out GeekoDoc v2 gradually and that
should help alleviate pain as well.